UPSTREAM: statx: Reserve the top bit of the mask for future struct expansion
Reserve the top bit of the mask for future expansion of the statx struct
and give an error if statx() sees it set.  All the other bits are ignored
if we see them set but don't support the bit; we just clear the bit in the
returned mask.

Change-Id: I5c42184bf771eb46227ba326e95ec1996b9ea1e5
Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: nostalgiceagle <183935403+nostalgiceagle@users.noreply.github.com>
(cherry picked from commit 11da3783c8c8f590e36cc59b79467bc8633ad00e)